requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.20
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4
gunicorn==21.2.0
psycopg2-binary==2.9.7  # Para PostgreSQL no Render
